# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "KServe推理服务管理后台API，提供InferenceService、LLMInferenceService、InferenceGraph等KServe资源的查询与管理功能",
        "title": "KServe Nexus API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5109",
    "basePath": "/",
    "paths": {
...
        "/kserve/llmisvc": {
            "get": {
                "description": "获取指定命名空间下的LLMInferenceService列表，若namespace=all则返回所有命名空间下的服务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "llmisvc"
                ],
                "summary": "获取LLM推理服务列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "命名空间，默认all表示所有命名空间",
                        "name": "namespace",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    },
                    "500": {
                        "description": "请求异常",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/kserve/llmisvc/{name}": {
            "get": {
                "description": "获取指定命名空间下单个LLMInferenceService的详细信息，包含部署拓扑图。kind=grafana时返回DOT格式，否则返回节点与边JSON",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "llmisvc"
                ],
                "summary": "获取单个LLM推理服务详情及拓扑图",
                "parameters": [
                    {
                        "type": "string",
                        "description": "LLMInferenceService名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "返回格式：grafana(DOT格式)或默认(节点边JSON)",
                        "name": "kind",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    },
                    "500": {
                        "description": "请求异常",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "middleware.Response": {
            "description": "返回通用信息",
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "数据内容"
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                }
            }
        }
    }
}